Migrate to AWS sdk v2: Lambda deployment (#1675)
**What this PR does / why we need it**:

**Which issue(s) this PR fixes**:

Ref #1663

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.

+++ pkg/app/piped/cloudprovider/lambda/function.go
@@ -56,8 +56,8 @@ type FunctionManifestSpec struct {
 	Name         string            `json:"name"`
 	Role         string            `json:"role"`
 	ImageURI     string            `json:"image"`
-	Memory       int64             `json:"memory"`
-	Timeout      int64             `json:"timeout"`
+	Memory       int32             `json:"memory"`
+	Timeout      int32             `json:"timeout"`
 	Tags         map[string]string `json:"tags,omitempty"`
 	Environments map[string]string `json:"environments,omitempty"`
 }
